Applied non-critical review fixes
