+++ ImageFeed/Profile/ProfilePresenterProtocol.swift
@@ -5,11 +5,10 @@
 //  Created by Pavel Seleznev on 3/17/25.
 //
 
-import UIKit
+import Foundation
 
 protocol ProfilePresenterProtocol {
     var view: ProfileViewControllerProtocol? { get set }
     
     func viewDidLoad()
-    func showLogoutConfirmationAlert() -> UIAlertController
 }
